using a tico.c file to include the header
I was using the core.c directly, now it uses a tico.c file just as
others nelua-decl bindings on the examples.

I also moved the running directory of gcc, it runs directly on the
repository's root instead of entering on nelua-decl.
