+++ gen-bindings.sh
@@ -2,16 +2,11 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-echo "gen-bindings: enter nelua-decl"
-cd nelua-decl
-
 echo "gen-bindings: running gcc to generate bindings"
 
-gcc -S ../tinycoffee/src/core.c \
-    -I ../tinycoffee/include -I ../tinycoffee/external \
-    -fplugin=./gcc-lua/gcc/gcclua.so \
-    -fplugin-arg-gcclua-script=../gen-scripts/tico.lua
-
-cd ..
+gcc -S ./tico.c \
+    -I ./tinycoffee/include/ -I ./tinycoffee/external/ \
+    -fplugin=./nelua-decl/gcc-lua/gcc/gcclua.so \
+    -fplugin-arg-gcclua-script=./tico.lua
 
 nelua basic-test.nelua --no-cache
