+++ setup-dev.sh
@@ -2,6 +2,10 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
+echo "setup-dev: removing possible existing tinycoffee and nelua-decl"
+rm -rf tinycoffee/
+rm -rf nelua-decl/
+
 echo "setup-dev: setup nelua-decl"
 
 # clone nelua-decl
