+++ tico.c
@@ -1,0 +1,1 @@
+#include <tico.h>
